ExternalizedMetadata tests split.
